Changes mostly as a result of __WXSTUBS__ compilation. The stubs code now compiles under Windows with VC++. Also OGL enhancements espec. wxDrawnShape.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@540 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/dc.cpp
@@ -386,7 +386,7 @@ void wxDC::DrawLine(long x1, long y1, long x2, long y2)
   CalcBoundingBox(x2, y2);
 }
 
-void wxDC::DrawArc(long x1,long y1,long x2,long y2,double xc,double yc)
+void wxDC::DrawArc(long x1,long y1,long x2,long y2, long xc, long yc)
 {
   double dx = xc-x1 ;
   double dy = yc-y1 ;
